# preppulse/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /focus-sessions:
    get:
      operationId: listFocusSessions
      tags: [study]
      summary: List focus sessions
      responses:
        "200":
          description: Focus sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FocusSession"
    post:
      operationId: createFocusSession
      tags: [study]
      summary: Start a timed focus session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FocusSessionInput"
      responses:
        "201":
          description: Started focus session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"
  /focus-sessions/{focusSessionId}:
    patch:
      operationId: updateFocusSession
      tags: [study]
      summary: Complete or abandon a focus session
      parameters:
        - name: focusSessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FocusSessionUpdate"
      responses:
        "200":
          description: Updated focus session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"
  /tasks:
    get:
      operationId: listTasks
      tags: [study]
      summary: List tasks
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [study]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [study]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [study]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    FocusSession:
      type: object
      properties:
        id:
          type: string
        subject:
          type: string
        plannedMinutes:
          type: number
        actualMinutes:
          type: number
        status:
          type: string
          enum: [active, completed, abandoned]
        taskId:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, subject, plannedMinutes, actualMinutes, status, taskId, startedAt, endedAt]
    FocusSessionInput:
      type: object
      properties:
        subject:
          type: string
          minLength: 1
        plannedMinutes:
          type: number
          minimum: 1
        taskId:
          type: string
          nullable: true
      required: [subject, plannedMinutes]
    FocusSessionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [completed, abandoned]
        actualMinutes:
          type: number
          minimum: 0
          maximum: 300
        taskId:
          type: string
          nullable: true
      required: [status, actualMinutes]
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, title, subject, status, createdAt, completedAt]
    TaskInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        subject:
          type: string
      required: [title]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        subject:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
...