# preppulse/lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { TaskUpdateStatus } from './taskUpdateStatus';

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  subject?: string;
  status?: TaskUpdateStatus;
}